Break out local visual test conf file
