+++ skyux-spa-visual-tests/config/shared.visual.conf.js
@@ -1,0 +1,17 @@
+/*jshint jasmine: true, node: true */
+'use strict';
+
+module.exports = {
+  useAllAngular2AppRoots: true,
+  beforeLaunch: function () {
+    require('ts-node').register({ ignore: false });
+  },
+
+  directConnect: true,
+  jasmineNodeOpts: {
+    showColors: true,
+    defaultTimeoutInterval: 30000
+  },
+  allScriptsTimeout: 11000,
+  specs: ['../src/app/**/*.visual-spec.ts']
+};
